separate build fetching
