+++ apps/builder/app/builder/features/sidebar-left/panels/marketplace/index.ts
@@ -1,1 +1,1 @@
-export * from "./tab";
+export * from "./marketplace";
