+++ apps/builder/app/shared/marketplace/types.ts
@@ -1,8 +1,13 @@
 import type { Project } from "@webstudio-is/prisma-client";
-import type { MarketplaceProduct } from "@webstudio-is/project-build";
+import type { Build, MarketplaceProduct } from "@webstudio-is/project-build";
 import type { Asset } from "@webstudio-is/sdk";
 
 export type MarketplaceOverviewItem = MarketplaceProduct & {
   projectId: Project["id"];
   thumbnailAssetName?: Asset["name"];
+};
+
+export type BuildData = {
+  build: Build;
+  assets: Array<[Asset["id"], Asset]>;
 };
